requirements.txt
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
rich==13.7.0
colorama==0.4.6
tabulate==0.9.0
